chore: changing GiftCard price to double

+++ common/src/main/java/org/dash/wallet/common/data/entity/GiftCard.kt
@@ -26,7 +26,7 @@ import org.bitcoinj.core.Sha256Hash
 data class GiftCard(
     @PrimaryKey var txId: Sha256Hash,
     var merchantName: String = "",
-    var price: Long = 0,
+    var price: Double = 0.0,
     var number: String? = null,
     var pin: String? = null,
     var barcodeValue: String? = null,
